+++ Frontend/src/features/inbox/pages/account-inbox/index.tsx
@@ -10,36 +10,20 @@ import SearchHeader from '@shared/components/inputs/SearchHeader';
 import Loader from '@shared/components/loader';
 import PaginationComponent from '@shared/components/table/Pagination';
 import FilterModal from '@shared/components/utils/FilterModal';
-import { DATE_RANGE_DROPDOWN_OPTIONS, UI_CONSTANTS } from '@shared/constants';
+import { UI_CONSTANTS } from '@shared/constants';
 import { useIsMobile } from '@shared/hooks';
-import { Filter, FilterOption, FilterOptionType } from '@shared/types';
+import { Filter } from '@shared/types';
 
 const AccountInboxPage: React.FC<{ account: string }> = ({ account }) => {
     const isMobile = useIsMobile();
 
     const {
         emails: { data: emailsData, fetchEmailsData, isLoadingEmails },
+        emailFilterOptions: { data: filterOptions },
         actions: { handleEmailSelect, handlePageSizeChange, handleResetPage, handleResetSelection },
         states: { selectedEmails, page, pageSize, searchValue, filter },
         setters: { setPage, setSearchValue, setFilter },
     } = useInboxPage(account);
-
-    const filterOptions: FilterOption[] = [
-        {
-            id: 1,
-            name: 'dateRange',
-            label: 'Date Range',
-            type: FilterOptionType.DROPDOWN,
-            data: DATE_RANGE_DROPDOWN_OPTIONS.map((item) => {
-                return {
-                    id: item.name,
-                    name: item.name,
-                    label: item.label,
-                    selectedValue: '',
-                };
-            }),
-        },
-    ];
 
     return (
         <>
@@ -47,7 +31,7 @@ const AccountInboxPage: React.FC<{ account: string }> = ({ account }) => {
                 <div className="flex h-full w-full flex-col items-center justify-center gap-4">
                     <APILoader show={isLoadingEmails} />
                     <div className="flex w-full gap-2">
-                        <FilterModal filter={filter} onFilterChange={(value: Filter) => setFilter(value)} filterOptions={filterOptions} />
+                        <FilterModal filter={filter} onFilterChange={(value: Filter) => setFilter(value)} filterOptions={filterOptions || []} />
                         <SearchHeader value={searchValue} onChange={setSearchValue} placeholder={UI_CONSTANTS.PLACEHOLDERS.SEARCH_EMAILS} />
                         <EmailMenuBarOptions
                             emailIds={selectedEmails}