+++ Frontend/src/shared/components/utils/FilterModal.tsx
@@ -8,6 +8,7 @@ import { Button } from '@shared/ui/button';
 import { Popover, PopoverContent, PopoverTrigger } from '@shared/ui/popover';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@shared/ui/select';
 import { ListFilter } from 'lucide-react';
+import { Switch } from '@shared/ui/switch';
 
 interface FilterModalProps {
     filter: Filter | null;
@@ -18,6 +19,8 @@ interface FilterModalProps {
 const FilterModal: React.FC<FilterModalProps> = ({ filter, onFilterChange, filterOptions }) => {
     const [localFilterObject, setLocalFilterObject] = useState<Filter | null>(filter || null);
     const [isOpen, setIsOpen] = useState(false);
+
+    console.log('localFilterObject', localFilterObject);
 
     useEffect(() => {
         setLocalFilterObject(filter);
@@ -37,11 +40,14 @@ const FilterModal: React.FC<FilterModalProps> = ({ filter, onFilterChange, filte
         return (val as string) || '';
     };
 
-    const handleValueChange = (name: string, value: string) => {
+    const handleValueChange = (name: string, value: string | boolean) => {
+        console.log('name', name, value);
         setLocalFilterObject((prev) => {
             const updated: Filter = { ...(prev || {}) };
-            if (name === 'accountId') {
+            if (name === 'accountId' && typeof value === 'string') {
                 updated.accountId = value ? [value] : undefined;
+            } else if (name === 'folders' && typeof value === 'string') {
+                updated.folders = value ? [value] : undefined;
             } else {
                 (updated as Record<string, unknown>)[name] = value;
             }
@@ -60,29 +66,44 @@ const FilterModal: React.FC<FilterModalProps> = ({ filter, onFilterChange, filte
                 <div className="flex w-full flex-col gap-2">
                     {filterOptions?.map((item) => {
                         return (
-                            <div className="flex w-full flex-col" key={item.id}>
-                                <p className="p-1 text-sm">{item.label}</p>
+                            <>
                                 {item.type === 'dropdown' && (
-                                    <Select value={getSelectedValue(item.name)} onValueChange={(value) => handleValueChange(item.name, value)}>
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue placeholder={`Select ${item.label}`} />
-                                        </SelectTrigger>
-                                        <SelectContent position="popper">
-                                            {item?.data &&
-                                                item?.data?.map((option, index) => {
-                                                    return (
-                                                        <SelectItem key={option.id || index} value={option.name} className="text-xs">
-                                                            {item.name === 'accountId' && option.provider && (
-                                                                <AccountProviderIcon provider={option.provider} className="size-4" />
-                                                            )}
-                                                            {option.label}
-                                                        </SelectItem>
-                                                    );
-                                                })}
-                                        </SelectContent>
-                                    </Select>
+                                    <div className="flex w-full flex-col" key={item.id}>
+                                        <p className="p-1 text-xs">{item.label}</p>
+                                        <Select value={getSelectedValue(item.name)} onValueChange={(value) => handleValueChange(item.name, value)}>
+                                            <SelectTrigger className="w-full text-xs">
+                                                <SelectValue placeholder={`Select ${item.label}`} />
+                                            </SelectTrigger>
+                                            <SelectContent position="popper" className="max-h-56 overflow-auto">
+                                                {Array.isArray(item?.data) &&
+                                                    item?.data?.length &&
+                                                    item?.data?.map((option, index) => {
+                                                        return (
+                                                            <SelectItem key={option.id || index} value={option.name} className="text-xs">
+                                                                {item.name === 'accountId' && option.provider && (
+                                                                    <AccountProviderIcon provider={option.provider} className="size-4" />
+                                                                )}
+                                                                {option.label}
+                                                            </SelectItem>
+                                                        );
+                                                    })}
+                                            </SelectContent>
+                                        </Select>
+                                    </div>
                                 )}
-                            </div>
+                                {item.type === 'toggle' && (
+                                    <>
+                                        <div className="flex items-center">
+                                            <p className="p-1 text-xs">{item.label}</p>
+                                            <Switch
+                                                id={item.name}
+                                                checked={localFilterObject?.unread}
+                                                onCheckedChange={(value) => handleValueChange(item.name, value)}
+                                            />
+                                        </div>
+                                    </>
+                                )}
+                            </>
                         );
                     })}
                 </div>